Improve services code
